requirements.txt
fastapi==0.115.9
uvicorn==0.34.2
pydantic==2.11.4
python-multipart==0.0.20
python-dotenv==1.1.0
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.59
langchain-cohere==0.4.4
cohere==5.15.0
mysql-connector-python==8.0.26
pandas==2.2.3
numpy==2.2.5
faiss-cpu==1.11.0
PyMySQL==1.1.1
